refactor: remove useless files
