+++ app/src/js/applications/structures/visualization/index.spec.js
